Add ignored users and clarify documentation
We recently added a bot to our github ecosystem that assigns issues and
pull_requests upon creation among other things. It has gotten really
annoying to be notified of actions by this bot. So now we can ignore it.

Also while I was in there I reworked the readme a touch to clarify the
installation instructions with links to the github/slack documentation
where applicable.