Run the build only on Push, not PR
For our purposes, the Push and PR run the same build because we always
require master to be merged into the branch before the PR can be merged.

Additionally drone secrets are not provided to PR builds by default
because a rogue PR could be created that would expose those secrets.

So, let's just turn of the PR build since we don't need it.